pyproject.toml
[project]
name = "ms2mol-evaluation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Marco Visani", email = "92217074+mvisani@users.noreply.github.com" },
]
requires-python = ">=3.12"
dependencies = [
    "cache-decorator>=2.2.0",
    "downloaders>=1.0.22",
    "dreams",
    "huggingface-hub>=0.33.0",
    "ipywidgets>=8.1.7",
    "matchms>=0.30.0",
    "matplotlib>=3.10.3",
    "matplotlib-venn>=1.1.2",
    "pandarallel>=1.6.5",
    "pandas>=2.3.0",
    "polars>=1.31.0",
    "psycopg2>=2.9.10",
    "pyarrow>=20.0.0",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
    "rdkit>=2024.9.5",
    "scikit-fingerprints>=1.16.0",
    "seaborn>=0.13.2",
    "tabulate>=0.9.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["ipykernel>=6.29.5", "jupyterlab>=4.4.3"]

[tool.uv.sources]
dreams = { git = "https://github.com/mvisani/DreaMS" }
